requirements.txt
hdbscan
hf-xet==1.1.2
inquirerpy
joblib
matplotlib
networkx
numba
numpy>=1.24
omegaconf
openai==1.82.0
pandas>=1.5
peft==0.11.1
pillow
PyYAML
pydantic
regex
requests
scikit-learn
scipy
sentence-transformers==4.1.0
sympy
tokenizers==0.21.1
torch==2.7.0
tqdm
transformers==4.52.3
umap-learn